# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderItem {
  productId: string;
  name: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  unitPrice: number;
  imageUrl?: string;
}

export interface CreateOrderRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  customerName: string;
  /**
     * @minLength 1
     * @maxLength 32
     */
  phone: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  address: string;
  /**
     * @minItems 1
     * @maxItems 50
     */
  items: OrderItem[];
}

export interface CreateOrderResponse {
  id: string;
  orderNumber: string;
  status: OrderStatus;
  total: number;
  createdAt: string;
}

export interface Order {
  id: string;
  orderNumber: string;
  customerName: string;
  phone: string;
  address: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  createdAt: string;
  updatedAt: string;
}

export interface AdminLoginRequest {
  /** @minLength 1 */
  password: string;
}

export interface AdminLoginResponse {
  token: string;
}

export interface AdminOrdersResponse {
  orders: Order[];
}

export interface AdminOrderResponse {
  order: Order;
}

export interface UpdateOrderStatusRequest {
  status: OrderStatus;
}

export interface ErrorResponse {
  message: string;
}

export type ListAdminOrdersParams = {
status?: OrderStatus;
};