# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Customer order operations
  - name: admin
    description: Admin order monitoring operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a customer order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Sign in to the order monitoring panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Admin session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List recent orders for the admin panel
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrdersResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}/status:
    patch:
      operationId: updateAdminOrderStatus
      tags: [admin]
      summary: Update an order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderResponse"
        "400":
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OrderStatus:
      type: string
      enum: [pending, confirmed, processing, shipped, delivered, cancelled]
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        quantity:
          type: number
          minimum: 1
        unitPrice:
          type: number
          minimum: 0
        imageUrl:
          type: string
      required: [productId, name, quantity, unitPrice]
    CreateOrderRequest:
      type: object
      properties:
        customerName:
          type: string
          minLength: 1
          maxLength: 120
        phone:
          type: string
          minLength: 1
          maxLength: 32
        address:
          type: string
          minLength: 1
          maxLength: 1000
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [customerName, phone, address, items]
    CreateOrderResponse:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        total:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, orderNumber, status, total, createdAt]
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        status:
          $ref: "#/components/schemas/OrderStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - customerName
        - phone
        - address
        - items
        - total
        - status
        - createdAt
        - updatedAt
    AdminLoginRequest:
      type: object
      properties:
        password:
          type: string
          minLength: 1
      required: [password]
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
      required: [token]
    AdminOrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required: [orders]
    AdminOrderResponse:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
      required: [order]
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
      required: [status]
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]