# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Create a customer order
 */
export const createOrderBodyCustomerNameMax = 120;

export const createOrderBodyPhoneMax = 32;

export const createOrderBodyAddressMax = 1000;


export const createOrderBodyItemsItemUnitPriceMin = 0;

export const createOrderBodyItemsMax = 50;



export const CreateOrderBody = zod.object({
  "customerName": zod.string().min(1).max(createOrderBodyCustomerNameMax),
  "phone": zod.string().min(1).max(createOrderBodyPhoneMax),
  "address": zod.string().min(1).max(createOrderBodyAddressMax),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "name": zod.string(),
  "quantity": zod.number().min(1),
  "unitPrice": zod.number().min(createOrderBodyItemsItemUnitPriceMin),
  "imageUrl": zod.string().optional()
})).min(1).max(createOrderBodyItemsMax)
})

export const CreateOrderResponse = zod.object({
  "id": zod.string(),
  "orderNumber": zod.string(),
  "status": zod.enum(['pending', 'confirmed', 'processing', 'shipped', 'delivered', 'cancelled']),
  "total": zod.number(),
  "createdAt": zod.coerce.date()
})


/**
 * @summary Sign in to the order monitoring panel
 */



export const AdminLoginBody = zod.object({
  "password": zod.string().min(1)
})

export const AdminLoginResponse = zod.object({
  "token": zod.string()
})


/**
 * @summary List recent orders for the admin panel
 */
export const ListAdminOrdersQueryParams = zod.object({
  "status": zod.enum(['pending', 'confirmed', 'processing', 'shipped', 'delivered', 'cancelled']).optional()
})


export const listAdminOrdersResponseOrdersItemItemsItemUnitPriceMin = 0;



export const ListAdminOrdersResponse = zod.object({
  "orders": zod.array(zod.object({
  "id": zod.string(),
  "orderNumber": zod.string(),
  "customerName": zod.string(),
  "phone": zod.string(),
  "address": zod.string(),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "name": zod.string(),
  "quantity": zod.number().min(1),
  "unitPrice": zod.number().min(listAdminOrdersResponseOrdersItemItemsItemUnitPriceMin),
  "imageUrl": zod.string().optional()
})),
  "total": zod.number(),
  "status": zod.enum(['pending', 'confirmed', 'processing', 'shipped', 'delivered', 'cancelled']),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
}))
})


/**
 * @summary Update an order status
 */
export const UpdateAdminOrderStatusParams = zod.object({
  "id": zod.coerce.string()
})

export const UpdateAdminOrderStatusBody = zod.object({
  "status": zod.enum(['pending', 'confirmed', 'processing', 'shipped', 'delivered', 'cancelled'])
})


export const updateAdminOrderStatusResponseOrderItemsItemUnitPriceMin = 0;



export const UpdateAdminOrderStatusResponse = zod.object({
  "order": zod.object({
  "id": zod.string(),
  "orderNumber": zod.string(),
  "customerName": zod.string(),
  "phone": zod.string(),
  "address": zod.string(),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "name": zod.string(),
  "quantity": zod.number().min(1),
  "unitPrice": zod.number().min(updateAdminOrderStatusResponseOrderItemsItemUnitPriceMin),
  "imageUrl": zod.string().optional()
})),
  "total": zod.number(),
  "status": zod.enum(['pending', 'confirmed', 'processing', 'shipped', 'delivered', 'cancelled']),
  "createdAt": zod.coerce.date(),
  "updatedAt": zod.coerce.date()
})
})