# lib/api-zod/src/generated/types/adminOrdersResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface AdminOrdersResponse {
  orders: Order[];
}