# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface CreateOrderRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  customerName: string;
  /**
     * @minLength 1
     * @maxLength 32
     */
  phone: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  address: string;
  /**
     * @minItems 1
     * @maxItems 50
     */
  items: OrderItem[];
}